initial commit
Project structure includes backend and frontend components. Backend is
the primary focus at the moment and needs to spin up to be a webserver
able to connect to the Stockfighter servers and create a local tracking
copy of data from the exchange while executing automated strategies
against it. The backend emits "monitor" signals for the frontend to, at
some point or another, display.